requirements.txt
python-telegram-bot
requests 
python-dotenv
flask
transformers
huggingface_hub
Pillow
boto3
openai
asyncio
functools